oss-agent-makers/adk-python/pyproject.toml
[project]
name = "adk-python"
version = "0.1.0"
description = "Tech Writer agent implemented using Google's Agent Development Kit (ADK)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk",
    "pathspec",
    "binaryornot",
    "openai",
    "litellm",
]
